refactor: segmentar youtube.ts en módulos organizados por recursos
- Crear estructura modular en src/lib/resources/
- YouTubeResource: clase base con lógica compartida (HTTP, retry, errores)
- VideosResource: endpoints de videos (getById, getMostPopular, etc.)
- ChannelsResource: endpoints de canales
- PlaylistsResource: endpoints de playlists
- SearchResource: búsquedas y videos relacionados
- Mantener 100% backward compatibility con API legacy
- Reducir youtube.ts de 753 a ~270 líneas
- Todos los tests pasan correctamente
