+++ src/lib/resources/index.ts
@@ -1,0 +1,5 @@
+export { YouTubeResource } from './base';
+export { VideosResource } from './videos';
+export { ChannelsResource } from './channels';
+export { PlaylistsResource } from './playlists';
+export { SearchResource } from './search';
